Up min Elixir version to 1.4, use GenServer, and clean ups
This ups the minimum version of Elixir to 1.4 because ExUnit.Formatter
started using GenServer for its formatters. In order to keep the
implementation simple, we won't support older versions. If there is
need to use versions lower than 1.4 of Elixir, then use version 1.3 of
this library.

Update travis for the new configuration.